+++ src/main/java/Language/Sentence.java
@@ -1,6 +1,8 @@
+package Language;
+
 import java.util.ArrayList;
 
-public final class Sentences {
+public final class Sentence {
     public static ArrayList<String> getSentences(){
         ArrayList<String> sentences = new ArrayList<>();
         sentences.add("1 is less than 5");
